Remove interfaces in model.text package
