+++ src/main/java/de/uhd/ifi/se/decision/management/jira/model/text/PartOfJiraIssueText.java
@@ -1,62 +1,252 @@
 package de.uhd.ifi.se.decision.management.jira.model.text;
 
-import org.codehaus.jackson.map.annotate.JsonDeserialize;
+import java.util.List;
 
-import com.atlassian.jira.issue.MutableIssue;
+import org.apache.commons.lang3.StringUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import com.atlassian.jira.component.ComponentAccessor;
+import com.atlassian.jira.issue.Issue;
+import com.atlassian.jira.issue.IssueManager;
+import com.atlassian.jira.issue.comments.Comment;
+import com.atlassian.jira.issue.comments.CommentManager;
 import com.atlassian.jira.issue.comments.MutableComment;
+import com.atlassian.jira.user.ApplicationUser;
 
-import de.uhd.ifi.se.decision.management.jira.model.text.impl.PartOfJiraIssueTextImpl;
+import de.uhd.ifi.se.decision.management.jira.model.DocumentationLocation;
+import de.uhd.ifi.se.decision.management.jira.model.KnowledgeElement;
+import de.uhd.ifi.se.decision.management.jira.model.impl.DecisionKnowledgeProjectImpl;
+import de.uhd.ifi.se.decision.management.jira.persistence.tables.PartOfJiraIssueTextInDatabase;
 
 /**
- * Interface for textual parts (substrings) of Jira issue comments or the
- * description. These parts can either be relevant decision knowledge elements
- * or irrelevant text.
+ * Models textual parts (substrings) of Jira issue comments or the description.
+ * These parts can either be relevant decision knowledge elements or irrelevant
+ * text.
  */
-@JsonDeserialize(as = PartOfJiraIssueTextImpl.class)
-public interface PartOfJiraIssueText extends PartOfText {
+public class PartOfJiraIssueText extends PartOfText {
+
+	private long commentId;
+
+	protected static final Logger LOGGER = LoggerFactory.getLogger(PartOfJiraIssueText.class);
+
+	public PartOfJiraIssueText() {
+		super();
+		this.documentationLocation = DocumentationLocation.JIRAISSUETEXT;
+	}
+
+	public PartOfJiraIssueText(PartOfText partOfText, Comment comment) {
+		this();
+		if (comment != null) {
+			this.setCommentId(comment.getId());
+			this.setJiraIssueId(comment.getIssue().getId());
+			this.setCreated(comment.getCreated());
+		}
+		this.setEndPosition(partOfText.getEndPosition());
+		this.setStartPosition(partOfText.getStartPosition());
+		this.setRelevant(partOfText.isRelevant());
+		this.setValidated(partOfText.isValidated());
+		this.setType(partOfText.getType());
+		this.setProject(partOfText.getProject());
+	}
+
+	public PartOfJiraIssueText(Comment comment) {
+		this(getFirstPartOfTextInComment(comment), comment);
+	}
+
+	private static PartOfText getFirstPartOfTextInComment(Comment comment) {
+		String projectKey = comment.getIssue().getProjectObject().getKey();
+		List<PartOfJiraIssueText> partsOfText = new TextSplitter().getPartsOfText(comment.getBody(), projectKey);
+		return partsOfText.get(0);
+	}
+
+	public PartOfJiraIssueText(PartOfText partOfText, Issue jiraIssue) {
+		this(partOfText, (Comment) null);
+		this.setCommentId(0);
+		this.setDescription(partOfText.getText());
+		this.setJiraIssueId(jiraIssue.getId());
+		this.setCreated(jiraIssue.getCreated());
+		this.setPlainText(!containsExcludedTag(partOfText.getText()));
+	}
+
+	public PartOfJiraIssueText(PartOfJiraIssueTextInDatabase databaseEntry) {
+		this(databaseEntry.getId(), databaseEntry.getEndPosition(), databaseEntry.getStartPosition(),
+				databaseEntry.isValidated(), databaseEntry.isRelevant(), databaseEntry.getProjectKey(),
+				databaseEntry.getCommentId(), databaseEntry.getJiraIssueId(), databaseEntry.getType(),
+				databaseEntry.getStatus());
+	}
+
+	public PartOfJiraIssueText(long id, int endSubstringCount, int startSubstringCount, boolean isValidated,
+			boolean isRelevant, String projectKey, long commentId, long issueId, String type, String status) {
+		this();
+		this.setId(id);
+		this.setEndPosition(endSubstringCount);
+		this.setStartPosition(startSubstringCount);
+		this.setValidated(isValidated);
+		this.setRelevant(isRelevant);
+		this.setProject(projectKey);
+		this.setCommentId(commentId);
+		this.setJiraIssueId(issueId);
+		this.setProject(new DecisionKnowledgeProjectImpl(projectKey));
+		this.setType(type);
+		this.setStatus(status);
+		Issue issue = ComponentAccessor.getIssueManager().getIssueObject(issueId);
+		if (issue != null) {
+			this.setKey(issue.getKey() + ":" + this.getId());
+		}
+		String text = "";
+		Comment comment = this.getComment();
+		if (comment == null) {
+			text = getJiraIssueDescription();
+			if (issue != null) {
+				this.setCreated(issue.getCreated());
+			}
+		} else {
+			text = comment.getBody();
+			this.setCreated(comment.getCreated());
+		}
+		try {
+			if (endSubstringCount < text.length()) {
+				text = text.substring(startSubstringCount, endSubstringCount);
+			} else if (endSubstringCount == text.length()) {
+				text = text.substring(startSubstringCount);
+			}
+		} catch (NullPointerException | StringIndexOutOfBoundsException e) {
+			LOGGER.error("Constructor faild to create object of PartOfJiraIssueText. Message: " + e.getMessage());
+		}
+		if (text == null) {
+			text = "";
+		}
+		this.setDescription(text);
+		this.setPlainText(!containsExcludedTag(text));
+		stripTagsFromBody(text);
+	}
+
+	public PartOfJiraIssueText(KnowledgeElement element) {
+		this.setId(element.getId());
+		this.setType(element.getType());
+		this.setSummary(element.getSummary());
+		this.setDescription(element.getDescription());
+		this.setProject(element.getProject());
+		this.setStatus(element.getStatus());
+	}
+
+	private boolean containsExcludedTag(String body) {
+		if (body == null) {
+			return false;
+		}
+		return StringUtils.indexOfAny(body.toLowerCase(), TextSplitter.EXCLUDED_TAGS) >= 0;
+	}
 
 	/**
-	 * Get the Jira issue description that the decision knowledge element or
-	 * irrelevant text is part of.
-	 * 
-	 * @see MutableIssue
-	 * @return Jira issue description as a String.
+	 * @return Jira issue comment as a {@link MutableComment} object that the
+	 *         decision knowledge element or irrelevant text is part of. Returns
+	 *         null if the part of text is within the description.
 	 */
-	String getJiraIssueDescription();
+	public MutableComment getComment() {
+		long commentId = getCommentId();
+		if (commentId <= 0) {
+			return null;
+		}
+		CommentManager commentManager = ComponentAccessor.getCommentManager();
+		if (commentManager == null) {
+			return null;
+		}
+		return commentManager.getMutableComment(commentId);
+	}
+
+	@Override
+	public String getText() {
+		Comment comment = this.getComment();
+		if (comment == null) {
+			return super.getSummary();
+		}
+		String body = comment.getBody().substring(this.getStartPosition(), this.getEndPosition());
+		return body.replaceAll("\\{.*?\\}", "");
+	}
+
+	@Override
+	public void setDescription(String body) {
+		super.setDescription(body);
+		super.setSummary(body);
+	}
+
+	@Override
+	public void setSummary(String body) {
+		super.setDescription(body);
+		super.setSummary(body);
+	}
+
+	private void stripTagsFromBody(String body) {
+		if (body == null) {
+			return;
+		}
+		String projectKey = this.getProject().getProjectKey();
+		if (TextSplitter.isAnyKnowledgeTypeTwiceExisting(body, projectKey)) {
+			int tagLength = 2 + TextSplitter.getKnowledgeTypeFromTag(body, projectKey).toString().length();
+			super.setDescription(body.substring(tagLength, body.length() - (tagLength)));
+			super.setSummary(super.getDescription());
+		} else {
+			super.setDescription(body.replaceAll("\\(.*?\\)", ""));
+			super.setSummary(super.getDescription());
+		}
+	}
 
 	/**
-	 * Set the id of the Jira issue comment that the decision knowledge element or
+	 * @return id of the Jira issue comment that the decision knowledge element or
+	 *         irrelevant text is part of. Returns 0 if the part of text is within
+	 *         the description.
+	 */
+	public long getCommentId() {
+		return this.commentId;
+	}
+
+	/**
+	 * Sets the id of the Jira issue comment that the decision knowledge element or
 	 * irrelevant text is part of.
 	 * 
 	 * @param id
 	 *            of the Jira issue comment.
 	 */
-	void setCommentId(long id);
+	public void setCommentId(long id) {
+		this.commentId = id;
+	}
 
 	/**
-	 * Get the id of the Jira issue comment that the decision knowledge element or
-	 * irrelevant text is part of.
-	 * 
-	 * @return id of the Jira issue comment. Returns 0 if the part of text is within
-	 *         the description.
+	 * @return Jira issue description that the decision knowledge element or
+	 *         irrelevant text is part of.
 	 */
-	long getCommentId();
+	public String getJiraIssueDescription() {
+		IssueManager issueManager = ComponentAccessor.getIssueManager();
+		Issue issue = issueManager.getIssueObject(this.getJiraIssueId());
+		if (issue == null) {
+			return super.getSummary();
+		}
+		return issue.getDescription();
+	}
+
+	@Override
+	public ApplicationUser getCreator() {
+		Comment comment = this.getComment();
+		if (comment == null) {
+			Issue issue = this.getJiraIssue();
+			return issue.getReporter();
+		}
+		return comment.getAuthorApplicationUser();
+	}
 
 	/**
-	 * Get the Jira issue comment that the decision knowledge element or irrelevant
-	 * text is part of.
-	 * 
-	 * @see MutableComment
-	 * @return Jira issue comment. Returns null if the part of text is within the
-	 *         description.
+	 * @return true if the textual parts (substrings) of Jira issue comments or the
+	 *         description is valid.
 	 */
-	MutableComment getComment();
-
-	/**
-	 * Determines whether the textual parts (substrings) of Jira issue comments or
-	 * the description is valid.
-	 * 
-	 * @return true if the object is valid.
-	 */
-	boolean isValid();
+	public boolean isValid() {
+		if (this.getEndPosition() == 0 && this.getStartPosition() == 0) {
+			return false;
+		}
+		if (this.getCommentId() <= 0) {
+			// documented in Jira issue description
+			return true;
+		}
+		return this.getComment() != null;
+	}
 }
